+++ mgt2e/module/documents/item.mjs
@@ -126,7 +126,14 @@ export class MgT2Item extends Item {
                     this.update({"system.weapon": item.system.weapon });
                     rollAttack(this.actor, item, skillDM);
                 } else if (item.system.weapon.magazine === 0) {
-                    rollAttack(this.actor, item, skillDM);
+                    if (this.hasTrait("oneUse")) {
+                        if (item.system.quantity > 0) {
+                            this.update({"system.quantity": item.system.quantity - 1});
+                            rollAttack(this.actor, item, skillDM);
+                        }
+                    } else {
+                        rollAttack(this.actor, item, skillDM);
+                    }
                 } else {
                     // No Ammo.
                 }
@@ -191,11 +198,21 @@ export class MgT2Item extends Item {
         return false;
     }
 
+    hasTrait(trait) {
+        if (this.type === "weapon" && this.system.weapon.traits) {
+            const regex = new RegExp(`(^|[, ])${trait}[^,]*($|[, ])`, 'gi');
+            return this.system.weapon.traits.match(regex) != null;
+        }
+        return false;
+    }
+
     useAmmo() {
         let weapon = this.system.weapon;
         if (weapon) {
             if (!isNaN(weapon.range) && parseInt(weapon.range) > 0) {
-                return true;
+                if (weapon.magazine > 0) {
+                    return true;
+                }
             }
             if (weapon.scale === "spacecraft") {
                 return true;